Fix case when session ID changes
We used to Add the new SessionId to the header, but when the SessionId
changed, header.Add didn't override the previous one. We used to just
add a new header that wasn't taken into account by the server, creating
thus a second HTTP Conflict
We know manually delete the previous SessionId to be sure that the good
one is in place.